Refactor(.feign): follow dto convention
