+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/client/KakaoUnlinkFeignClient.java
@@ -8,14 +8,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 
+import com.newbarams.ajaja.infra.feign.kakao.model.KakaoRequest;
 import com.newbarams.ajaja.infra.feign.kakao.model.KakaoResponse;
-import com.newbarams.ajaja.infra.feign.kakao.model.KakaoUnlinkRequest;
 
 @FeignClient(name = "KakaoDisconnectFeignClient", url = "https://kapi.kakao.com")
 public interface KakaoUnlinkFeignClient {
 	@PostMapping(value = "/v1/user/unlink", consumes = APPLICATION_FORM_URLENCODED_VALUE)
 	KakaoResponse.Withdraw unlink(
 		@RequestHeader(AUTHORIZATION) String adminKey,
-		@RequestBody KakaoUnlinkRequest request
+		@RequestBody KakaoRequest.Unlink request
 	);
 }
