--- src/main/java/com/newbarams/ajaja/infra/feign/kakao/model/KakaoTokenRequest.java
@@ -1,19 +1,0 @@
-package com.newbarams.ajaja.infra.feign.kakao.model;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class KakaoTokenRequest {
-	private static final String KAKAO_GRANT_TYPE = "authorization_code";
-
-	private String grant_type;
-	private String client_id;
-	private String redirect_uri;
-	private String code;
-	private String client_secret;
-
-	public KakaoTokenRequest(String clientId, String redirectUri, String code, String clientSecret) {
-		this(KAKAO_GRANT_TYPE, clientId, redirectUri, code, clientSecret);
-	}
-}
