--- src/main/java/com/newbarams/ajaja/infra/feign/kakao/model/KakaoUnlinkRequest.java
@@ -1,16 +1,0 @@
-package com.newbarams.ajaja.infra.feign.kakao.model;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class KakaoUnlinkRequest {
-	private static final String KAKAO_TARGET_ID_TYPE = "user_id";
-
-	private String target_id_type;
-	private Long target_id;
-
-	public KakaoUnlinkRequest(Long targetId) {
-		this(KAKAO_TARGET_ID_TYPE, targetId);
-	}
-}
