+++ src/main/java/com/newbarams/ajaja/infra/feign/ncp/client/NaverSendAlimtalkFeignClient.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import com.newbarams.ajaja.infra.feign.ncp.model.NaverAlimtalkRequest;
+import com.newbarams.ajaja.infra.feign.ncp.model.NaverRequest;
 import com.newbarams.ajaja.infra.feign.ncp.model.NaverResponse;
 
 @FeignClient(
@@ -17,5 +17,5 @@ import com.newbarams.ajaja.infra.feign.ncp.model.NaverResponse;
 )
 public interface NaverSendAlimtalkFeignClient {
 	@PostMapping(value = "/alimtalk/v2/services/{id}/messages", consumes = APPLICATION_JSON_VALUE)
-	NaverResponse.AlimTalk send(@PathVariable("id") String serviceId, @RequestBody NaverAlimtalkRequest request);
+	NaverResponse.AlimTalk send(@PathVariable("id") String serviceId, @RequestBody NaverRequest.Alimtalk request);
 }
