+++ src/test/java/com/newbarams/ajaja/infra/feign/ncp/client/NaverSendAlimtalkFeignClientTest.java
@@ -1,5 +1,6 @@
 package com.newbarams.ajaja.infra.feign.ncp.client;
 
+import static com.newbarams.ajaja.infra.feign.ncp.model.NaverRequest.*;
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Disabled;
@@ -7,7 +8,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.newbarams.ajaja.infra.feign.ncp.model.NaverAlimtalkRequest;
 import com.newbarams.ajaja.infra.feign.ncp.model.NaverResponse;
 
 @Disabled
@@ -21,11 +21,11 @@ class NaverSendAlimtalkFeignClientTest {
 	@Test
 	void send_Success_WithRealApiCall() {
 		// given
-		NaverAlimtalkRequest request =
-			NaverAlimtalkRequest.remind("01012345678", "테스트임당", "안녕하세요?", "https://www.ajaja.me/plans/1");
+		Alimtalk request = Alimtalk.remind("01012345678", "테스트임당", "안녕하세요?", "https://www.ajaja.me/plans/1");
 
 		// when
-		NaverResponse.AlimTalk response = naverSendAlimtalkFeignClient.send(naverCloudProperties.getServiceId(), request);
+		NaverResponse.AlimTalk response =
+			naverSendAlimtalkFeignClient.send(naverCloudProperties.getServiceId(), request);
 
 		// then
 		assertThat(response).isNotNull();
